rename file in src
